+++ src/components/cards/Patient Cards/PatientHorizontal.js
@@ -7,13 +7,15 @@ const PatientHorizontal = (props) => {
 
     const dispatch = useDispatch()
 
-    const visitPatientProfile=()=>{
-        window.location.href = "/clinic/patient"; 
+    const visitPatientProfile = () => {
+        window.location.href = "/clinic/patient";
     }
 
-    const deletePatient=()=>{
-        // alert("You are about to delete patient")
-        dispatch(deletePatientAction(props.patient))
+    const deletePatient = () => {
+        var result = window.confirm("Are you sure, you want to delete?");
+        if (result) {
+            dispatch(deletePatientAction(props.patient))
+        }
     }
 
     return (
@@ -24,8 +26,8 @@ const PatientHorizontal = (props) => {
             <div className="col text-primary">{props.patient.patient_id}</div>
             <div className="col">{props.patient.name}</div>
             <div className="col">
-            <div className='btn btn-success m-1'><Link className='text-white text-decoration-none' to={`/clinic/patient?id=${props.patient.patient_id}`}>Visit</Link></div>
-                <div className='btn btn-danger m-1' onClick={()=>{deletePatient()}}>delete</div>
+                <div className='btn btn-success m-1'><Link className='text-white text-decoration-none' to={`/clinic/patient?id=${props.patient.patient_id}`}>Visit</Link></div>
+                <div className='btn btn-danger m-1' onClick={() => { deletePatient() }}>delete</div>
             </div>
         </div>
     )
